Added Frame Limit for optimalization and Map generation rewrite
Had a lot of work done on the map generation and the camera system.
Firstly I made the bounding boxes of the camera check if they hit the border of the Map,
So it doesnt render anything that we shouldnt see.
Nextly I also added bound checking for the player too, so the player cant go out of bounds
There is still a bug with it, where the player cant move closer to the border, still have to figure that out
And lastly i added a framelimit, right now its set for 999 because, its a lightweight game,
so a modern pc can spit out 999 frames easily but, ill make a menu for that too.
And I added deltaTime into consideration so the player movement speed doesnt collerate whatsoever to the
framerate. Thats about it!

Signed-off-by: Krisztián Hozák <business.frimex@gmail.com>